requirements.txt
fastapi
uvicorn
python-dotenv
python-docx
google-generativeai
pdfplumber
docx2txt